backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
python-dotenv>=1.0.1
pillow>=10.4.0
numpy>=1.26.4
opencv-python>=4.10.0.84
invisible-watermark>=0.2.0
reedsolo>=1.7.0
cryptography>=43.0.0
pdqhash>=0.2.6
google-genai>=1.0.0